Issue #7: Support attachment of RAW disks
Refactored much of the disk processing parts.

Now we will try to set a unique UUID. If that fails, we will treat the disk
as a raw disk and try to convert it to VDI format. Exception made to
cdrom disks. For cdrom disks we create a .iso symlink so VirtualBox
does not complain when attaching them. Of course this means we only support
ISO images for the moment, not CDR or DMG or RAW for the cdrom part.